+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 1.2.0 *(unreleased)*
 ----------------------------
 
 * Support using Gradle Build Cache with Yarn tasks (#205) _(mark-vieira)_
+* Bumped Gradle wrapper version to 3.4.1
 * ...
 
 Version 1.1.1 *(2017-01-16)*