+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Version 1.2.0 *(unreleased)*
 ----------------------------
 
 * Support using Gradle Build Cache with Yarn tasks (#205) _(mark-vieira)_
-* Bumped Gradle wrapper version to 3.4.1
+* Bumped Gradle wrapper version to 3.5
 * New args can be added for npmSetup and yarnSetup (#226)
 * Uses --no-save for npmSetup and yarnSetup (#222)
 * Added execOverrides and ignoreExitValue on npmSetupTask (#196)