+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.2.0 *(unreleased)*
 * Uses --no-save for npmSetup and yarnSetup (#222)
 * Added execOverrides and ignoreExitValue on npmSetupTask (#196)
 * Added gruntFile in grunt extension to use different file (#189)
+* npm_* tasks using nodeModulesDir (##136) _(janario)_
 * ...
 
 Version 1.1.1 *(2017-01-16)*