+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Version 1.2.0 *(unreleased)*
+Version 1.2.0 *(2017-06-14)*
 ----------------------------
 
 * Support using Gradle Build Cache with Yarn tasks (#205) _(mark-vieira)_
@@ -11,7 +11,6 @@ Version 1.2.0 *(unreleased)*
 * Added execOverrides and ignoreExitValue on npmSetupTask (#196)
 * Added gruntFile in grunt extension to use different file (#189)
 * npm_* tasks using nodeModulesDir (##136) _(janario)_
-* ...
 
 Version 1.1.1 *(2017-01-16)*
 ----------------------------